Houston trip/requirements.txt
fpdf==1.7.2
markdown==3.4.1
beautifulsoup4==4.11.1
requests==2.28.1
Pillow==9.3.0
reportlab==3.6.12
